+++ geoextent/__init__.py
@@ -1,3 +1,3 @@
 name = "geoextent"
 
-__version__ = '0.5.2'
+__version__ = '0.6.0'